# src/main/kotlin/com/jiubredeemer/rulebook/domain/race/dto/RaceDto.java
package com.jiubredeemer.rulebook.domain.race.dto;

import lombok.Data;

import java.util.UUID;

@Data
public class RaceDto {
    private UUID id;
    private UUID roomId;
    private Boolean hidden;
    private String name;
    private String description;
    private String code;
    private String speciesCode;
    private String imgUrl;
    private RaceStatsDto stats;
}